Use the bucket API from hashbrown v0.16.1
This is similar to indexmap-rs/indexmap#424, simplify a lot of code by
using a bucket index in `OccupiedEntry` rather than having to split
references. The effects are a little more pronounced here because our
`RefMut` needed **3** references, due to the extra offset tracking, so
with that gone, now *every* `*Entry` type is reduced in size.